+++ app/src/main/java/joshuatee/wx/settings/Location.kt
@@ -26,10 +26,10 @@ import joshuatee.wx.objects.LatLon
 import joshuatee.wx.util.Utility
 import java.util.Locale
 import joshuatee.wx.common.GlobalDictionaries
-import joshuatee.wx.getHtml
 import joshuatee.wx.parse
 import joshuatee.wx.radar.RadarSites
 import joshuatee.wx.util.To
+import joshuatee.wx.util.UtilityDownloadNws
 import joshuatee.wx.util.UtilityString
 import joshuatee.wx.util.WfoSites
 
@@ -130,8 +130,9 @@ object Location {
     }
 
     private fun getWfoRadarSiteFromPoint(location: LatLon): List<String> {
-        val pointData =
-            ("https://api.weather.gov/points/" + location.latString + "," + location.lonString).getHtml()
+//        val pointData =
+//            ("https://api.weather.gov/points/" + location.latForNws + "," + location.lonForNws).getHtml()
+        val pointData = UtilityDownloadNws.getLocationPointData(location)
         // "cwa": "IWX",
         // "radarStation": "KGRR"
         val wfo = pointData.parse("\"cwa\": \"(.*?)\"")