+++ doc/ChangeLog.md
@@ -2,9 +2,9 @@
 
 # Developer ChangeLog
 
-## 55951 2025_03_27
+## 55951 2025_04_04
 
-* [ADD]
+* [ADD] use UtilityDownloadNws.getLocationPointData in Location.getWfoRadarSiteFromPoint
 
 ## 55950 2025_03_27